requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
PyYAML>=6.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
rcon>=2.3.0

